+++ lib/analysis/simulateTRAN.ts
@@ -1,17 +1,136 @@
 import { EPS } from "../constants/EPS"
+import { VT_300K } from "../constants/physics"
 import { solveReal } from "../math/solveReal"
 import type { ParsedCircuit } from "../parsing/parseNetlist"
 import { stampAdmittanceReal } from "../stamping/stampAdmittanceReal"
 import { stampCurrentReal } from "../stamping/stampCurrentReal"
 import { stampVoltageSourceReal } from "../stamping/stampVoltageSourceReal"
 
+/**
+ * Compute a stable timestep and number of steps for the transient analysis.
+ * - If a non-zero dt is provided, use that.
+ * - Otherwise, default to 1000 steps across the stop time.
+ */
+function computeEffectiveTimeStep(dtRequested: number, tstop: number) {
+  const dtEff = dtRequested > EPS ? dtRequested : Math.max(tstop / 1000, EPS)
+  const steps = Math.max(1, Math.ceil(tstop / Math.max(dtEff, EPS)))
+  const dt = steps > 0 ? tstop / steps : tstop
+  return { dt, steps }
+}
+
+/**
+ * Stamps all elements (R, C, L, switches, voltage sources, diodes) into A and b
+ * for the given time t and Newton iteration.
+ */
+function stampAllElementsAtTime(
+  A: number[][],
+  b: number[],
+  ckt: ParsedCircuit,
+  t: number,
+  dt: number,
+  x: number[],
+  iter: number,
+) {
+  // Resistors (conductance to ground and between nodes)
+  for (const r of ckt.R) {
+    const G = 1 / r.R
+    stampAdmittanceReal(A, ckt.nodes, r.n1, r.n2, G)
+  }
+
+  // Capacitors via Thevenin companion: Gc and Ieq
+  for (const c of ckt.C) {
+    const Gc = c.C / Math.max(dt, EPS)
+    stampAdmittanceReal(A, ckt.nodes, c.n1, c.n2, Gc)
+    const Ieq = -Gc * c.vPrev
+    stampCurrentReal(b, ckt.nodes, c.n1, c.n2, Ieq)
+  }
+
+  // Inductors via Norton companion: Gl and current source
+  for (const l of ckt.L) {
+    const Gl = Math.max(dt, EPS) / l.L
+    stampAdmittanceReal(A, ckt.nodes, l.n1, l.n2, Gl)
+    stampCurrentReal(b, ckt.nodes, l.n1, l.n2, l.iPrev)
+  }
+
+  // Voltage-controlled switches as conductances based on last state
+  for (const sw of ckt.S) {
+    const model = sw.model
+    if (!model) continue
+    const Rvalue = sw.isOn ? model.Ron : model.Roff
+    const Rclamped = Math.max(Math.abs(Rvalue), EPS)
+    const G = 1 / Rclamped
+    stampAdmittanceReal(A, ckt.nodes, sw.n1, sw.n2, G)
+  }
+
+  // Independent voltage sources (DC or waveform)
+  for (const vs of ckt.V) {
+    const Vt = vs.waveform ? vs.waveform(t) : vs.dc || 0
+    stampVoltageSourceReal(A, b, ckt.nodes, vs, Vt)
+  }
+
+  // Diodes (Shockley model) using companion linearization for Newton step
+  for (const d of ckt.D) {
+    const model = d.model
+    if (!model) continue
+
+    const { nPlus, nMinus } = d
+
+    const vp_idx = ckt.nodes.matrixIndexOfNode(nPlus)
+    const vn_idx = ckt.nodes.matrixIndexOfNode(nMinus)
+
+    const v_plus_prev_iter = nPlus === 0 ? 0 : (x[vp_idx] ?? 0)
+    const v_minus_prev_iter = nMinus === 0 ? 0 : (x[vn_idx] ?? 0)
+    const vd_prev_iter = v_plus_prev_iter - v_minus_prev_iter
+
+    const vd = iter === 0 ? d.vdPrev : vd_prev_iter
+
+    // Diode equation with Shockley model, using companion model for NR
+    const v_thermal = model.N * VT_300K
+    let vd_limited = vd
+    if (vd > 0.8) vd_limited = 0.8 // prevent overflow
+    if (vd < -1.0) vd_limited = -1.0 // limit reverse voltage
+
+    const exp_val = Math.exp(vd_limited / v_thermal)
+    const id = model.Is * (exp_val - 1)
+    const gd = Math.max((model.Is / v_thermal) * exp_val, 1e-12)
+
+    const ieq = id - gd * vd_limited
+
+    stampAdmittanceReal(A, ckt.nodes, nPlus, nMinus, gd)
+    stampCurrentReal(b, ckt.nodes, nPlus, nMinus, ieq)
+  }
+}
+
+/**
+ * Update switch states based on the latest solution vector.
+ * Returns true if any switch changed state (causing another NR iteration).
+ */
+function updateSwitchStatesFromSolution(ckt: ParsedCircuit, x: number[]) {
+  let switched = false
+  for (const sw of ckt.S) {
+    const model = sw.model
+    if (!model) continue
+    const vp = sw.ncPos === 0 ? 0 : (x[sw.ncPos - 1] ?? 0)
+    const vn = sw.ncNeg === 0 ? 0 : (x[sw.ncNeg - 1] ?? 0)
+    const vctrl = vp - vn
+    let nextState = sw.isOn
+    if (sw.isOn) {
+      if (vctrl < model.Voff) nextState = false
+    } else if (vctrl > model.Von) {
+      nextState = true
+    }
+    if (nextState !== sw.isOn) {
+      sw.isOn = nextState
+      switched = true
+    }
+  }
+  return switched
+}
+
 function simulateTRAN(ckt: ParsedCircuit) {
   if (!ckt.analyses.tran) return null
   const { dt: dtRequested, tstop } = ckt.analyses.tran
-  const effectiveDt =
-    dtRequested > EPS ? dtRequested : Math.max(tstop / 1000, EPS)
-  const steps = Math.max(1, Math.ceil(tstop / Math.max(effectiveDt, EPS)))
-  const dt = steps > 0 ? tstop / steps : tstop
+  const { dt, steps } = computeEffectiveTimeStep(dtRequested, tstop)
 
   const nNodeVars = ckt.nodes.count() - 1
   const nVsrc = ckt.V.length
@@ -33,92 +152,11 @@ function simulateTRAN(ckt: ParsedCircuit) {
       const A = Array.from({ length: Nvar }, () => new Array(Nvar).fill(0))
       const b = new Array(Nvar).fill(0)
 
-      for (const r of ckt.R) {
-        const G = 1 / r.R
-        stampAdmittanceReal(A, ckt.nodes, r.n1, r.n2, G)
-      }
-
-      for (const c of ckt.C) {
-        const Gc = c.C / Math.max(dt, EPS)
-        stampAdmittanceReal(A, ckt.nodes, c.n1, c.n2, Gc)
-        const Ieq = -Gc * c.vPrev
-        stampCurrentReal(b, ckt.nodes, c.n1, c.n2, Ieq)
-      }
-
-      for (const l of ckt.L) {
-        const Gl = Math.max(dt, EPS) / l.L
-        stampAdmittanceReal(A, ckt.nodes, l.n1, l.n2, Gl)
-        stampCurrentReal(b, ckt.nodes, l.n1, l.n2, l.iPrev)
-      }
-
-      for (const sw of ckt.S) {
-        const model = sw.model
-        if (!model) continue
-        const Rvalue = sw.isOn ? model.Ron : model.Roff
-        const Rclamped = Math.max(Math.abs(Rvalue), EPS)
-        const G = 1 / Rclamped
-        stampAdmittanceReal(A, ckt.nodes, sw.n1, sw.n2, G)
-      }
-
-      for (const vs of ckt.V) {
-        const Vt = vs.waveform ? vs.waveform(t) : vs.dc || 0
-        stampVoltageSourceReal(A, b, ckt.nodes, vs, Vt)
-      }
-
-      // Diode companion model stamping
-      const VT = 0.02585 // Thermal voltage at 300K
-      for (const d of ckt.D) {
-        const model = d.model
-        if (!model) continue
-
-        const { nPlus, nMinus } = d
-
-        const vp_idx = ckt.nodes.matrixIndexOfNode(nPlus)
-        const vn_idx = ckt.nodes.matrixIndexOfNode(nMinus)
-
-        const v_plus_prev_iter = nPlus === 0 ? 0 : (x[vp_idx] ?? 0)
-        const v_minus_prev_iter = nMinus === 0 ? 0 : (x[vn_idx] ?? 0)
-        const vd_prev_iter = v_plus_prev_iter - v_minus_prev_iter
-
-        const vd = iter === 0 ? d.vdPrev : vd_prev_iter
-
-        // Diode equation with Shockley model, using companion model for NR
-        const v_thermal = model.N * VT
-        let vd_limited = vd
-        if (vd > 0.8) vd_limited = 0.8 // prevent overflow
-        if (vd < -1.0) vd_limited = -1.0 // limit reverse voltage
-
-        const exp_val = Math.exp(vd_limited / v_thermal)
-        const id = model.Is * (exp_val - 1)
-        const gd = Math.max((model.Is / v_thermal) * exp_val, 1e-12)
-
-        const ieq = id - gd * vd_limited
-
-        stampAdmittanceReal(A, ckt.nodes, nPlus, nMinus, gd)
-        stampCurrentReal(b, ckt.nodes, nPlus, nMinus, ieq)
-      }
+      stampAllElementsAtTime(A, b, ckt, t, dt, x, iter)
 
       x = solveReal(A, b)
 
-      let switched = false
-      for (const sw of ckt.S) {
-        const model = sw.model
-        if (!model) continue
-        const vp = sw.ncPos === 0 ? 0 : (x[sw.ncPos - 1] ?? 0)
-        const vn = sw.ncNeg === 0 ? 0 : (x[sw.ncNeg - 1] ?? 0)
-        const vctrl = vp - vn
-        let nextState = sw.isOn
-        if (sw.isOn) {
-          if (vctrl < model.Voff) nextState = false
-        } else if (vctrl > model.Von) {
-          nextState = true
-        }
-        if (nextState !== sw.isOn) {
-          sw.isOn = nextState
-          switched = true
-        }
-      }
-
+      const switched = updateSwitchStatesFromSolution(ckt, x)
       if (!switched) break
       if (iter === 19) break
     }
@@ -174,8 +212,7 @@ function simulateTRAN(ckt: ParsedCircuit) {
       const v2 = nMinus === 0 ? 0 : (x[nMinus - 1] ?? 0)
       const vd = v1 - v2
 
-      const VT = 0.02585
-      const v_thermal = model.N * VT
+      const v_thermal = model.N * VT_300K
       const exp_val = Math.exp(vd / v_thermal)
       const id = model.Is * (exp_val - 1)
       ;(elementCurrents[d.name] ||= []).push(id)
