+++ lib/constants/physics.ts
@@ -1,0 +1,1 @@
+export const VT_300K = 0.02585
