+++ lib/parsing/parseNetlist.ts
@@ -1,4 +1,7 @@
 import { EPS } from "../constants/EPS"
+import { NodeIndex } from "./NodeIndex"
+import { parseNumberWithUnits } from "./units"
+import { parsePulseArgs, parsePwlArgs, pulseValue, pwlValue } from "./waveforms"
 
 type Waveform = ((t: number) => number) | null
 
@@ -108,71 +111,7 @@ type ParsedCircuit = {
   }
 }
 
-class NodeIndex {
-  private map: Map<string, number>
-  rev: string[]
-
-  constructor() {
-    this.map = new Map([["0", 0]])
-    this.rev = ["0"]
-  }
-
-  getOrCreate(name: string) {
-    const origName = String(name)
-    const key = origName.toUpperCase()
-    if (this.map.has(key)) return this.map.get(key) as number
-    const idx = this.rev.length
-    this.map.set(key, idx)
-    this.rev.push(origName)
-    return idx
-  }
-
-  get(name: string) {
-    return this.map.get(String(name).toUpperCase())
-  }
-
-  count() {
-    return this.rev.length
-  }
-
-  matrixIndexOfNode(nodeId: number) {
-    if (nodeId === 0) return -1
-    return nodeId - 1
-  }
-}
-
 type CircuitNodeIndex = NodeIndex
-
-function parseNumberWithUnits(raw: unknown) {
-  if (raw == null) return NaN
-  let s = String(raw).trim()
-  if (s === "") return NaN
-  if (/^[+-]?\d*\.?\d+(?:[eE][+-]?\d+)?$/.test(s)) return parseFloat(s)
-  const unitMul = {
-    t: 1e12,
-    g: 1e9,
-    meg: 1e6,
-    k: 1e3,
-    m: 1e-3,
-    u: 1e-6,
-    n: 1e-9,
-    p: 1e-12,
-    f: 1e-15,
-  } as const
-  const m = s.match(/^([+-]?\d*\.?\d+(?:[eE][+-]?\d+)?)([a-zA-Z]+)$/)
-  if (!m) return parseFloat(s)
-  const [, numberPart, suffixPart] = m
-  if (numberPart == null) return parseFloat(s)
-  let val = parseFloat(numberPart)
-  let suf = (suffixPart ?? "").toLowerCase()
-  suf = suf.replace(/(ohm|v|a|s|h|f)$/g, "")
-  if (suf === "meg") return val * unitMul.meg
-  if (suf.length === 1 && suf in unitMul) {
-    const key = suf as keyof typeof unitMul
-    return val * unitMul[key]
-  }
-  return val
-}
 
 function smartTokens(line: string) {
   const re = /"[^"]*"|\w+\s*\([^)]*\)|\([^()]*\)|\S+/g
@@ -186,78 +125,6 @@ function requireToken(tokens: string[], index: number, context: string) {
   const token = tokens[index]
   if (token == null) throw new Error(context)
   return token
-}
-
-function parsePulseArgs(token: string): PulseSpec {
-  const clean = token.trim().replace(/^pulse\s*\(/i, "(")
-  const inside = clean.replace(/^\(/, "").replace(/\)$/, "").trim()
-  const parts = inside.split(/[\s,]+/).filter((x) => x.length)
-  if (parts.length < 7) throw new Error("PULSE(...) requires 7 or 8 args")
-  const vals = parts.map((value) => parseNumberWithUnits(value))
-  if (vals.some((v) => Number.isNaN(v)))
-    throw new Error("Invalid PULSE() numeric value")
-  return {
-    v1: vals[0]!,
-    v2: vals[1]!,
-    td: vals[2]!,
-    tr: vals[3]!,
-    tf: vals[4]!,
-    ton: vals[5]!,
-    period: vals[6]!,
-    ncycles: parts[7] != null ? vals[7]! : Infinity,
-  }
-}
-
-function parsePwlArgs(token: string) {
-  const clean = token.trim().replace(/^pwl\s*\(/i, "(")
-  const inside = clean.replace(/^\(/, "").replace(/\)$/, "").trim()
-  const parts = inside.split(/[\s,]+/).filter((x) => x.length)
-  if (parts.length === 0 || parts.length % 2 !== 0)
-    throw new Error("PWL(...) requires an even number of time/value pairs")
-  const pairs: { t: number; v: number }[] = []
-  for (let i = 0; i < parts.length; i += 2) {
-    const t = parseNumberWithUnits(parts[i]!)
-    const v = parseNumberWithUnits(parts[i + 1]!)
-    if (Number.isNaN(t) || Number.isNaN(v))
-      throw new Error("Invalid PWL() numeric value")
-    pairs.push({ t, v })
-  }
-  return pairs
-}
-
-function pwlValue(pairs: { t: number; v: number }[], t: number): number {
-  if (pairs.length === 0) return 0
-  if (t <= pairs[0]!.t) return pairs[0]!.v
-  for (let i = 1; i < pairs.length; i++) {
-    const prev = pairs[i - 1]!
-    const curr = pairs[i]!
-    if (t <= curr.t) {
-      const dt = Math.max(curr.t - prev.t, EPS)
-      const a = (t - prev.t) / dt
-      return prev.v + (curr.v - prev.v) * a
-    }
-  }
-  return pairs[pairs.length - 1]!.v
-}
-
-function pulseValue(p: PulseSpec, t: number) {
-  if (t < p.td) return p.v1
-  const tt = t - p.td
-  const cyclesDone = Math.floor(tt / p.period)
-  if (cyclesDone >= p.ncycles) return p.v1
-  const tc = tt - cyclesDone * p.period
-  if (tc < p.tr) {
-    const a = tc / Math.max(p.tr, EPS)
-    return p.v1 + (p.v2 - p.v1) * a
-  }
-  if (tc < p.tr + p.ton) {
-    return p.v2
-  }
-  if (tc < p.tr + p.ton + p.tf) {
-    const a = (tc - (p.tr + p.ton)) / Math.max(p.tf, EPS)
-    return p.v2 + (p.v1 - p.v2) * a
-  }
-  return p.v1
 }
 
 function parseNetlist(text: string): ParsedCircuit {
